requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-tinymce==3.4.0
django-versatileimagefield==2.2
Pillow==9.2.0
python-decouple==3.6
python-magic==0.4.27
sqlparse==0.4.2
